Patch test for input validation

+++ CIS227-cpp-Group4-Assignment1/Calculations.cpp
@@ -7,7 +7,7 @@
 #include "Calculations.h"
 
 //function to calculate hypotenuse
-float hypot(float a, float b)
+double hypot(double a, double b)
 {
 	return sqrt((pow(a, 2) + pow(b, 2)));
 }
